# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample API for demonstration purposes.",
        "title": "My API",
        "contact": {},
        "version": "1.0"
    },
    "host": "8.218.102.112:4880",
    "basePath": "/api/v1",
    "paths": {
...
        "/chains/:chainType/gas_fee": {
            "get": {
                "description": "Get the current gas fee for Solana blockchain",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "on-chain-data"
                ],
                "summary": "Get Solana gas fee",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Type of the token (e.g., sol, eth)",
                        "name": "chainType",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/httpRespone.GasFee"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/chains/:chainType/sol_price": {
            "get": {
                "description": "Get the current real-time price for Solana",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "on-chain-data"
                ],
                "summary": "Get Solana real-time price",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Type of the token (e.g., sol, eth)",
                        "name": "chainType",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "number"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/token_balance/{chainType}/{owner}/{token}": {
            "get": {
                "description": "Get the balance of a token for a given owner",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "on-chain-data"
                ],
                "summary": "Get token balance",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Type of the token (e.g., sol, eth)",
                        "name": "chainType",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Owner of the token",
                        "name": "owner",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Token address",
                        "name": "token",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.TokenBalance"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/token_pool_info_sol/{chainType}/{tokenAddress}": {
            "get": {
                "description": "Retrieve pool information for a specific token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pool"
                ],
                "summary": "Get token pool information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Type of the token (e.g., sol)",
                        "name": "chainType",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Address of the token",
                        "name": "tokenAddress",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.MarketInfo"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/tokens/{klineType}/{chainType}/{tokenAddress}": {
            "get": {
                "description": "Get historical kline (candlestick) data for a specific token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "on-chain-data"
                ],
                "summary": "Get token kline data",
                "parameters": [
                    {
                        "enum": [
                            "kline",
                            "mcapkline"
                        ],
                        "type": "string",
                        "description": "Kline type",
                        "name": "klineType",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "sol",
                            "eth",
                            "btc"
                        ],
                        "type": "string",
                        "description": "Chain type",
                        "name": "chainType",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Token address",
                        "name": "tokenAddress",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Resolution (1S=1sec, 1=1min, 5=5min, 15=15min, 60=1hour, 240=4hour, 720=12hour, 1D=1day)",
                        "name": "resolution",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/response.KlineData"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid parameters",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "httpRespone.GasFee": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/httpRespone.GasFeeResponse"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "httpRespone.GasFeeResponse": {
            "type": "object",
            "properties": {
                "priorityFeeLevels": {
                    "$ref": "#/definitions/httpRespone.PriorityFeeLevels"
                }
            }
        },
        "httpRespone.PriorityFeeLevels": {
            "type": "object",
            "properties": {
                "high": {
                    "type": "number"
                },
                "low": {
                    "type": "number"
                },
                "medium": {
                    "type": "number"
                },
                "min": {
                    "type": "number"
                },
                "unsafeMax": {
                    "type": "number"
                },
                "veryHigh": {
                    "type": "number"
                }
            }
        },
...
        "response.KlineData": {
            "type": "object",
            "properties": {
                "close": {
                    "type": "number"
                },
                "high": {
                    "type": "number"
                },
                "low": {
                    "type": "number"
                },
                "open": {
                    "type": "number"
                },
                "timestamp": {
                    "type": "integer"
                },
                "volume": {
                    "type": "number"
                }
            }
        },
        "response.MarketInfo": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "base_address": {
                    "type": "string"
                },
                "base_reserve": {
                    "type": "integer"
                },
                "base_reserve_value": {
                    "type": "number"
                },
                "base_symbol": {
                    "type": "string"
                },
                "base_vault_address": {
                    "type": "string"
                },
                "circulating_supply": {
                    "type": "number"
                },
                "creation_timestamp": {
                    "type": "integer"
                },
                "creator": {
                    "type": "string"
                },
                "initial_base_reserve": {
                    "type": "integer"
                },
                "initial_liquidity": {
                    "type": "number"
                },
                "initial_quote_reserve": {
                    "type": "integer"
                },
                "liquidity": {
                    "type": "number"
                },
                "market_cap": {
                    "type": "number"
                },
                "pool_address": {
                    "type": "string"
                },
                "quote_address": {
                    "type": "string"
                },
                "quote_reserve": {
                    "type": "integer"
                },
                "quote_reserve_value": {
                    "type": "number"
                },
                "quote_symbol": {
                    "type": "string"
                },
                "quote_vault_address": {
                    "type": "string"
                },
                "total_supply": {
                    "type": "number"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "error": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "response.TokenBalance": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "string"
                },
                "owner": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
...
    }
}